Step 2 - With Database/requirements.txt
Flask==2.0.1
Jinja2==3.0.3
requests==2.25.1
psycopg2-binary